Remove specific community health files
This removes the files from the repository so that they can be inherited
from the EasyDynamics/.github repository. In the future, if we need
OSCAL React Library-specific files, we can always add them again. For
now, this allows us to leverage the shared language on security,
contributing, and the code of conduct.